requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-RESTful==0.3.10
Flask-Swagger-UI==4.11.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
mysqlclient==2.2.1
redis==5.0.1
celery==5.3.6
line-bot-sdk==3.7.0
pandas==2.1.4
openpyxl==3.1.2
pytest==7.4.3
python-dateutil==2.8.2
gunicorn==21.2.0

# Security
Flask-Limiter==3.5.0
cryptography==41.0.5
Flask-Audit==1.0.0

# Performance
Flask-Caching==2.1.0

# Monitoring & Logging
prometheus-flask-exporter==0.22.4
sentry-sdk==1.32.0

# Testing
pytest-cov==4.1.0
selenium==4.15.2

# i18n
Flask-Babel==4.0.0

# Backup
python-crontab==3.0.0
